+++ tests/testthat/test-utils.R
@@ -1,3 +1,0 @@
-test_that("testing works", {
-  succeed("Remove this test once real tests exist.")
-})